backend/requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==46.0.7
fastapi==0.115.0
google-ai-generativelanguage==0.6.10
google-api-core==2.25.2
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.3
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.27.2
idna==3.13
JsonForm==0.0.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
JsonSir==0.0.2
numpy==2.4.4
pandas==3.0.2
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
Python-EasyConfig==0.1.7
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
Resource==0.2.1
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.30.6
